Fixed player id error in read_utils.c
Fixed an error where identification of player was incorrect when player fails to place a valid coordinate. Was previously looking for the 6th element in a line that looked like: "<got (X): [12, 0]" or "<got (O): [5, 14]" but this fails when a line like: "Player with O: error on input" is given. Switching to using ft_strchr and searching for 'O' should hopefully be more robust.